Moved Cuisine Model to separate file
